refactor: helpers.ts moved to utils folder
